Refactor out deployment resize function
