+++ plugins/knife/knife-deployment-resize.rb
@@ -33,14 +33,15 @@
 # as those licenses appear in the file LICENSE-OPENSSL.
 
 require_relative 'knife-clearwater-utils'
+require_relative 'knife-deployment-utils'
 require_relative 'trigger-chef-client'
 require_relative 'knife-shared-config-update'
-require_relative 'boxes'
 
 module ClearwaterKnifePlugins
   class DeploymentResize < Chef::Knife
     include ClearwaterKnifePlugins::ClearwaterUtils
     include ClearwaterKnifePlugins::TriggerChefClient
+    include ClearwaterKnifePlugins::DeploymentUtils
 
     banner "knife deployment resize -E ENV"
 
@@ -50,18 +51,8 @@ module ClearwaterKnifePlugins
       require 'nokogiri'
       require 'parallel'
       require 'chef/knife/client_delete'
-      require_relative 'knife-box-create'
-      require_relative 'knife-box-delete'
-      require_relative 'knife-deployment-clean'
-      require_relative 'knife-security-groups-create'
-      require_relative 'knife-dnszone-create'
-      require_relative 'knife-dns-records-create'
-      require_relative 'knife-bind-records-create'
       require_relative 'dns-records'
-      BoxCreate.load_deps
-      DeploymentClean.load_deps
       DnsRecordsCreate.load_deps
-      BindRecordsCreate.load_deps
     end
 
     %w{bono homestead homer ibcf sprout sipp ralf}.each do |node|
@@ -126,154 +117,6 @@ module ClearwaterKnifePlugins
     # Estimated number of busy hour calls per subscriber.
     BHCA_PER_SUB = 2
 
-    def launch_box(box, environment, retries)
-      success = false
-
-      # Since we run this in an aggressively multi-threaded way, smear our start
-      # times out randomly over a 5 second period to avoid spamming cloud
-      # provisioning APIs.
-      sleep(rand * 5)
-
-      loop do
-        begin
-          box_create = BoxCreate.new("-E #{environment}".split)
-          box_create.name_args = [box[:role]]
-          box_create.config[:index] = box[:index]
-          box_create.config[:verbosity] = config[:verbosity]
-          Chef::Config[:verbosity] = config[:verbosity]
-          box_create.config[:cloud] = config[:cloud]
-          box_create.config[:seagull] = config[:seagull]
-          box_create.config[:ralf] = (config[:ralf_count] and (config[:ralf_count] > 0))
-          box_create.run
-        rescue Exception => e
-          Chef::Log.error "Failed to create node: #{e}"
-          Chef::Log.debug e.backtrace
-        end
-
-        box_name = node_name_from_definition(environment, box[:role], box[:index])
-        Chef::Log.debug "Checking successful creation of #{box_name}"
-        begin
-          node = Chef::Node.load(box_name)
-          if node.roles.include? box[:role]
-            Chef::Log.info "Successfully created #{box_name}"
-            break
-          else
-            Chef::Log.error "Failed to set roles for #{box_name}"
-            delete_box(box_name, environment)
-          end
-        rescue
-          Chef::Log.error "Failed to create node for #{box_name}"
-          clean_up_broken_client(box_name, environment)
-          @fail_count += 1
-        end
-
-        # Bail out if we've hit too many failures across the worker threads
-        return false if @fail_count >= retries
-      end
-
-      return true
-    end
-
-    def clean_up_broken_client(box_name, environment)
-      client = find_clients(name: box_name)
-      client.each do
-        client_delete = Chef::Knife::ClientDelete.new
-        client_delete.name_args = [box_name]
-        client_delete.config[:yes] = true
-        client_delete.config[:verbosity] = config[:verbosity]
-        client_delete.run
-      end
-    end
-
-    def delete_box(box_name, env)
-      box_delete = BoxDelete.new("-E #{env}".split)
-      box_delete.name_args = [box_name]
-      box_delete.config[:yes] = true
-      box_delete.config[:purge] = true
-      box_delete.config[:verbosity] = config[:verbosity]
-      Chef::Config[:verbosity] = config[:verbosity]
-      box_delete.run(true)
-    end
-
-    def launch_boxes(box_list)
-      @fail_count = 0
-      results = Parallel.map(box_list, in_threads: box_list.length) do |box|
-        if @fail_count < config[:fail_limit]
-          launch_box(box, config[:environment], config[:fail_limit])
-        else
-          false
-        end
-      end
-
-      abort_deployment if results.any? { |r| not r }
-    end
-
-    def potential_deletions
-      victims = find_nodes(roles: "clearwater-infrastructure")
-      # Only delete nodes with roles contained in this whitelist
-      whitelist = ["bono", "ellis", "ibcf", "homer", "homestead", "sprout", "sipp", "ralf", "seagull"]
-      victims.select! { |v| not (v.roles & whitelist).empty? }
-      # Don't delete any AIO/AMI nodes
-      victims.delete_if { |v| v.roles.include? "cw_aio" }
-      return victims
-    end
-
-    def in_stable_state? env
-      transitioning_list = find_quiescing_nodes(env)
-      return transitioning_list.empty?
-    end
-
-    def prepare_to_quiesce_extra_boxes(env, orig_box_list)
-      victims = potential_deletions
-      box_list = orig_box_list.map { |b| node_name_from_definition(env, b[:role], b[:index]) }
-
-      victims.select! { |v| not box_list.include? v.name }
-
-      return if victims.empty?
-
-      victims.each do |v|
-        prepare_to_quiesce_box(v.name, env)
-      end
-    end
-
-    def quiesce_extra_boxes(env, box_list)
-      victims = potential_deletions
-      box_list.map! { |b| node_name_from_definition(env, b[:role], b[:index]) }
-
-      victims.select! { |v| not box_list.include? v.name }
-
-      return if victims.empty?
-
-      victims.each do |v|
-        quiesce_box(v.name, env)
-      end
-    end
-
-    def delete_quiesced_boxes(env)
-      record_manager = Clearwater::DnsRecordManager.new(attributes["root_domain"])
-
-      quiesced_boxes = find_quiescing_nodes(env)
-      record_manager.delete_node_records(quiesced_boxes)
-
-      quiesced_boxes.each do |v|
-        delete_box(v.name, env)
-      end
-    end
-
-    def calculate_boxes_to_create(env, nodes)
-      current_nodes = find_nodes(roles: "clearwater-infrastructure")
-
-      result = nodes.select do |node|
-        not current_nodes.any? { |cnode| cnode.name == node_name_from_definition(env, node[:role], node[:index]) }
-      end
-
-      return result
-    end
-
-    def node_name_from_definition(environment, role, index)
-      "#{environment}-#{role}-#{index}"
-    end
-
     def get_current_counts
       result = Hash.new(0)
       %w{bono ellis ibcf homer homestead sprout sipp ralf seagull}.each do |node|
@@ -310,31 +153,6 @@ module ClearwaterKnifePlugins
       end
     end
 
-    def confirm_changes(old, new)
-      # Don't touch any AIO or AMI nodes
-      old_names = potential_deletions.map {|v| v.name}
-      new_names = create_cluster(new).map do |n|
-        node_name_from_definition(env, n[:role], n[:index])
-      end
-      create_boxes = new_names - old_names
-      victim_boxes = old_names - new_names
-
-      unless create_boxes.empty?
-        ui.msg "The following boxes will be created:"
-        create_boxes.each do |b|
-          ui.msg " - #{b}"
-        end
-      end
-      unless victim_boxes.empty?
-        ui.msg "The following boxes will be quiesced:"
-        victim_boxes.each do |b|
-          ui.msg " - #{b}"
-        end
-      end
-
-      fail "Exiting on user request" unless continue?
-    end
-
     def calculate_box_counts(config)
       Chef::Log.info "Subscriber count given, calculating box counts automatically:"
 
@@ -358,16 +176,10 @@ module ClearwaterKnifePlugins
       calculate_box_counts(config) if config[:subscribers]
 
       # Initialize status object
-      init_status
+      init_status(["bono", "ellis", "homer", "homestead", "sprout", "sipp", "ralf"], ["seagull"])
 
       # Create security groups
-      status["Security Groups"][:status] = "Configuring..."
-      Chef::Log.info "Creating security groups..."
-      sg_create = SecurityGroupsCreate.new("-E #{config[:environment]}".split)
-      sg_create.config[:verbosity] = config[:verbosity]
-      Chef::Config[:verbosity] = config[:verbosity]
-      sg_create.run
-      status["Security Groups"][:status] = "Done"
+      configure_security_groups(config, SecurityGroupsCreate)
       set_progress 10
 
       # Enumerate current box counts so we can compare the desired list
@@ -390,7 +202,8 @@ module ClearwaterKnifePlugins
         seagull: seagull_count || old_counts[:seagull] }
 
       # Confirm changes if there are any
-      confirm_changes(old_counts, new_counts) unless old_counts == new_counts
+      whitelist = ["bono", "ellis", "ibcf", "homer", "homestead", "sprout", "sipp", "ralf", "seagull"]
+      confirm_changes(old_counts, new_counts, whitelist) unless old_counts == new_counts
 
       # Create boxes
       node_list = create_cluster(new_counts)
@@ -400,25 +213,20 @@ module ClearwaterKnifePlugins
       launch_boxes(create_node_list)
       set_progress 50
 
-      prepare_to_quiesce_extra_boxes(env.name, node_list)
+      prepare_to_quiesce_extra_boxes(env.name, node_list, whitelist)
 
       if not in_stable_state? env
         Chef::Log.info "Removing nodes from DNS before quiescing..."
-        configure_dns config
+        configure_dns(config, DnsRecordsCreate)
         Chef::Log.info "Waiting 60s for DNS to propagate..."
         sleep 60
       end
 
-      quiesce_extra_boxes(env.name, node_list)
+      quiesce_extra_boxes(env.name, node_list, whitelist)
       set_progress 60
 
       # Now that all the boxes are in place, cleanup any that failed
-      Chef::Log.info "Cleaning deployment..."
-      deployment_clean = DeploymentClean.new("-E #{config[:environment]}".split)
-      deployment_clean.config[:verbosity] = config[:verbosity]
-      deployment_clean.config[:cloud] = config[:cloud]
-      Chef::Config[:verbosity] = config[:verbosity]
-      deployment_clean.run(yes_allowed=true)
+      clean_deployment config
       set_progress 70
 
       # Sleep to let chef catch up _sigh_
@@ -468,17 +276,11 @@ module ClearwaterKnifePlugins
 
       Chef::Log.info "Sleeping for 90 seconds before updating DNS to allow cluster to synchronize..."
       sleep(90)
-      # Setup DNS zone record
-      status["DNS"][:status] = "Configuring..."
-      Chef::Log.info "Creating zone record..."
-      zone_create = DnszoneCreate.new
-      zone_create.config[:verbosity] = config[:verbosity]
-      Chef::Config[:verbosity] = config[:verbosity]
-      zone_create.name_args = [attributes["root_domain"]]
-      zone_create.run
-      set_progress 95
 
-      configure_dns config
+      # Setup DNS zone record
+      configure_dns_zone(config, attributes)
+      set_progress 95
+      configure_dns(config, DnsRecordsCreate)
       set_progress 99
 
       Chef::Log.info "Finishing resize operation"
@@ -495,65 +297,6 @@ module ClearwaterKnifePlugins
         Chef::Log.info "Applying shared configuration..."
         SharedConfigUpdate.new("-E #{config[:environment]}".split).run
       end
-    end
-
-    def configure_dns config
-      # Setup DNS records defined above
-      if config[:cloud].to_sym == :openstack
-        Chef::Log.info "Creating BIND records..."
-        bind_create = BindRecordsCreate.new("-E #{config[:environment]}".split)
-        bind_create.config[:verbosity] = config[:verbosity]
-        Chef::Config[:verbosity] = config[:verbosity]
-        bind_create.run
-        status["DNS"][:status] = "Done"
-      else
-        Chef::Log.info "Creating DNS records..."
-        dns_create = DnsRecordsCreate.new("-E #{config[:environment]}".split)
-        dns_create.config[:verbosity] = config[:verbosity]
-        Chef::Config[:verbosity] = config[:verbosity]
-        dns_create.run
-        status["DNS"][:status] = "Done"
-      end
-    end
-
-    # Expands out hashes of boxes, e.g. {:bono => 3} becomes:
-    # {{:role => "bono", :index => 1}, {:role => "bono", :index = 2}, etc...
-    def create_cluster(boxes)
-      boxes.map {|box, n| (1..n).map {|i| {:role => box.to_s, :index => i}}}.flatten
-    end
-
-    def init_status
-      Thread.current[:progress] = 0
-      Thread.current[:status] = {"Nodes" => {}}
-      ["Security Groups", "DNS"].each do |item|
-        Thread.current[:status][item] = {:status => "Pending"}
-      end
-
-      ["bono", "ellis", "homer", "homestead", "sprout", "sipp", "ralf"].each do |node|
-        Thread.current[:status]["Nodes"][node] =
-          {:status => "Pending", :count => config["#{node}_count".to_sym]}
-      end
-      ["seagull"].each do |node|
-        Thread.current[:status]["Nodes"][node] =
-          {:status => "Pending", :count => (config[:seagull] ? 1 : 0)}
-      end
-    end
-
-    def set_progress(pct)
-      Thread.current[:progress] = pct
-    end
-
-    def status
-      Thread.current[:status]
-    end
-
-    def abort_deployment
-      msg = "Too many failures (#{config[:fail_limit]}), aborting...
-      To clean up broken boxes in deployment, issue:
-      knife deployment clean -E #{config[:environment]}
-      To delete the deployment completely, issue:
-      knife deployment delete -E #{config[:environment]}"
-      fail msg
     end
   end
 end
