+++ roles/ralf.rb
@@ -35,6 +35,7 @@
 name "ralf"
 description "ralf role"
 run_list [
+  "recipe[clearwater::local_config]",
   "role[clearwater-infrastructure]",
   "recipe[clearwater::ralf]",
   "role[clearwater-etcd]"
