+++ roles/seagull.rb
@@ -35,6 +35,7 @@
 name "seagull"
 description "seagull role"
 run_list [
+  "recipe[clearwater::local_config]",
   "role[clearwater-infrastructure]",
   "recipe[clearwater::seagull]"
 ]
